optimize msg object
