+++ service/common/util.go
@@ -60,52 +60,3 @@ func ParseMessage(enablePackage bool, ctx *gin.Context, packager *ecdsa.Packager
 	}
 	return msgObject, nil
 }
-
-func getMsgObjectByType(messageType MessageType) (interface{}, error) {
-	var req interface{}
-	switch messageType {
-	case InvitationType:
-		req = &message.Invitation{}
-	case ConnectionRequestType:
-		req = &message.ConnectionRequest{}
-	case ConnectionResponseType:
-		req = &message.ConnectionResponse{}
-	case ConnectionAckType:
-		req = &message.ConnectionACK{}
-	case DisconnectType, SendDisconnectType:
-		req = &message.DisconnectRequest{}
-	case SendProposalCredentialType:
-		req = &message.ProposalCredential{}
-	case OfferCredentialType:
-		req = &message.OfferCredential{}
-	case ProposalCredentialType:
-		req = &message.ProposalCredential{}
-	case SendRequestCredentialType:
-		req = &message.RequestCredential{}
-	case RequestCredentialType:
-		req = &message.RequestCredential{}
-	case IssueCredentialType:
-		req = &message.IssueCredential{}
-	case CredentialAckType:
-		req = &message.CredentialACK{}
-	case RequestPresentationType:
-		req = &message.RequestPresentation{}
-	case SendRequestPresentationType:
-		req = &message.RequestPresentation{}
-	case PresentationType:
-		req = &message.Presentation{}
-	case PresentationAckType:
-		req = &message.PresentationACK{}
-	case SendBasicMsgType:
-		req = &message.BasicMessage{}
-	case QueryBasicMessageType:
-		req = &message.QueryGeneralMessageRequest{}
-	case QueryCredentialType:
-		req = &message.QueryCredentialRequest{}
-	case QueryPresentationType:
-		req = &message.QueryPresentationRequest{}
-	default:
-		return nil, fmt.Errorf("msg type err:%v", messageType)
-	}
-	return req, nil
-}
